Fix issues from moving components around.
